Add stuff to env.js
