+++ next.config.mjs
@@ -5,7 +5,7 @@ import withSerwistInit from "@serwist/next";
  * Run `build` or `dev` with `SKIP_ENV_VALIDATION` to skip env validation.
  * This is especially useful for Docker builds.
  */
-await import("./src/env.mjs");
+await import("./src/env.js");
 
 /** @type {import('next').NextConfig} */
 const config = {
