+++ src/env.js
@@ -1,0 +1,52 @@
+import { createEnv } from "@t3-oss/env-nextjs";
+import { z } from "zod";
+
+export const env = createEnv({
+	/*
+	 * Serverside Environment variables, not available on the client.
+	 * Will throw if you access these variables on the client.
+	 */
+	server: {
+		DATABASE_URL: z.string().url(),
+		NODE_ENV: z
+			.enum(["development", "test", "production"])
+			.default("development"),
+		NEXTAUTH_SECRET:
+			process.env.NODE_ENV === "production"
+				? z.string()
+				: z.string().optional(),
+		NEXTAUTH_URL: z.preprocess(
+			// This makes Vercel deployments not fail if you don't set NEXTAUTH_URL
+			// Since NextAuth.js automatically uses the VERCEL_URL if present.
+			(str) => process.env.VERCEL_URL ?? str,
+			// VERCEL_URL doesn't include `https` so it cant be validated as a URL
+			process.env.VERCEL ? z.string() : z.string().url(),
+		),
+		GOOGLE_ID: z.string(),
+		GOOGLE_SECRET: z.string(),
+		SECRET_KEY: z.string(),
+	},
+	/*
+	 * Environment variables available on the client (and server).
+	 *
+	 * 💡 You'll get type errors if these are not prefixed with NEXT_PUBLIC_.
+	 */
+	client: {
+		NEXT_PUBLIC_SITE_URL: z.string().url(),
+	},
+
+	// For Next.js >= 13.4.4, you only need to destructure client variables:
+	experimental__runtimeEnv: {
+		NEXT_PUBLIC_SITE_URL: process.env.NEXT_PUBLIC_SITE_URL,
+	},
+	/**
+	 * Run `build` or `dev` with `SKIP_ENV_VALIDATION` to skip env validation. This is especially
+	 * useful for Docker builds.
+	 */
+	skipValidation: !!process.env.SKIP_ENV_VALIDATION,
+	/**
+	 * Makes it so that empty strings are treated as undefined. `SOME_VAR: z.string()` and
+	 * `SOME_VAR=''` will throw an error.
+	 */
+	emptyStringAsUndefined: true,
+});
