--- src/env.mjs
@@ -1,34 +1,0 @@
-import { createEnv } from "@t3-oss/env-nextjs";
-import { z } from "zod";
-
-export const env = createEnv({
-	server: {
-		DATABASE_URL: z.string().url(),
-		NODE_ENV: z.enum(["development", "test", "production"]),
-		NEXTAUTH_SECRET: z.string().min(1),
-		NEXTAUTH_URL: z.preprocess(
-			// This makes Vercel deployments not fail if you don't set NEXTAUTH_URL
-			// Since NextAuth.js automatically uses the VERCEL_URL if present.
-			(str) => process.env.VERCEL_URL ?? str,
-			// VERCEL_URL doesn't include `https` so it cant be validated as a URL
-			process.env.VERCEL ? z.string() : z.string().url(),
-		),
-		GOOGLE_ID: z.string(),
-		GOOGLE_SECRET: z.string(),
-		SECRET_KEY: z.string(),
-	},
-	client: {
-		NEXT_PUBLIC_SITE_URL: z.string().url(),
-	},
-
-	runtimeEnv: {
-		DATABASE_URL: process.env.DATABASE_URL,
-		NODE_ENV: process.env.NODE_ENV,
-		NEXTAUTH_SECRET: process.env.NEXTAUTH_SECRET,
-		NEXTAUTH_URL: process.env.NEXTAUTH_URL,
-		GOOGLE_ID: process.env.GOOGLE_ID,
-		GOOGLE_SECRET: process.env.GOOGLE_SECRET,
-		SECRET_KEY: process.env.SECRET_KEY,
-		NEXT_PUBLIC_SITE_URL: process.env.NEXT_PUBLIC_SITE_URL,
-	},
-});
