+++ tsconfig.json
@@ -36,6 +36,7 @@
 		"**/*.tsx",
 		"**/*.cjs",
 		"**/*.mjs",
+		"**/*.js",
 		".next/types/**/*.ts"
 	],
 	"exclude": ["node_modules"]
